feat: add HDCL HOLLAR deploy scripts (10% APY, 1M facilitator)
Deploy HOLLAR aToken, stable/variable debt tokens, and interest rate
strategy for the HDCL market instance.

Run with: MARKET_NAME=HDCL npx hardhat deploy --tags hdcl_hollar_deploy
