+++ deploy/15_deploy_hdcl_hollar_stable_debt.ts
@@ -1,7 +1,6 @@
 import { DeployFunction } from 'hardhat-deploy/types';
 import { getPool } from '@galacticcouncil/aave-deploy-v3/dist/helpers/contract-getters';
 import { ZERO_ADDRESS } from '../helpers/constants';
-import { GhoAToken } from '../types';
 
 const func: DeployFunction = async function ({ getNamedAccounts, deployments, ...hre }) {
   const { deploy } = deployments;
@@ -9,34 +8,30 @@ const func: DeployFunction = async function ({ getNamedAccounts, deployments, ..
 
   const pool = await getPool();
 
-  const aTokenResult = await deploy('GhoAToken-HDCL', {
+  const stableDebtResult = await deploy('GhoStableDebtToken-HDCL', {
     from: deployer,
-    contract: 'GhoAToken',
+    contract: 'GhoStableDebtToken',
     args: [pool.address],
     log: true,
     gasLimit: 10_000_000,
   });
-  const aTokenImpl = (await hre.ethers.getContractAt(
-    'GhoAToken',
-    aTokenResult.address
-  )) as GhoAToken;
-  const initializeTx = await aTokenImpl.initialize(
+  const stableDebtImpl = await hre.ethers.getContract('GhoStableDebtToken-HDCL');
+  const initializeTx = await stableDebtImpl.initialize(
     pool.address, // initializingPool
-    ZERO_ADDRESS, // treasury
     ZERO_ADDRESS, // underlyingAsset
     ZERO_ADDRESS, // incentivesController
-    0, // aTokenDecimals
-    'GHO_ATOKEN_HDCL_IMPL', // aTokenName
-    'GHO_ATOKEN_HDCL_IMPL', // aTokenSymbol
-    '0x10' // params
+    0, // debtTokenDecimals
+    'HOLLAR_STABLE_DEBT_HDCL_IMPL', // debtTokenName
+    'HOLLAR_STABLE_DEBT_HDCL_IMPL', // debtTokenSymbol
+    0 // params
   );
   await initializeTx.wait();
 
-  console.log(`GhoAToken-HDCL Implementation: ${aTokenResult.address}`);
+  console.log(`GhoStableDebtToken-HDCL Implementation: ${stableDebtResult.address}`);
   return true;
 };
 
-func.id = 'GhoAToken-HDCL';
-func.tags = ['GhoAToken-HDCL', 'hdcl_gho_deploy'];
+func.id = 'GhoStableDebt-HDCL';
+func.tags = ['GhoStableDebt-HDCL', 'hdcl_hollar_deploy'];
 
 export default func;
